Move CLI into separate file
